Moved edit tag to filters section
