+++ app/ui/modal-provider.tsx
@@ -13,18 +13,15 @@ import useProject from "#/lib/swr/use-project";
 import Cookies from "js-cookie";
 import { useAddProjectModal } from "@/components/app/modals/add-project-modal";
 import { useAcceptInviteModal } from "@/components/app/modals/accept-invite-modal";
-import { useTagLinkModal } from "@/components/app/modals/tag-link-modal";
 import { useAddEditDomainModal } from "@/components/app/modals/add-edit-domain-modal";
 import { useGoogleOauthModal } from "@/components/app/modals/google-oauth-modal";
 
 export const ModalContext = createContext<{
   setShowAddProjectModal: Dispatch<SetStateAction<boolean>>;
   setShowAddEditDomainModal: Dispatch<SetStateAction<boolean>>;
-  setShowTagLinkModal: Dispatch<SetStateAction<boolean>>;
 }>({
   setShowAddProjectModal: () => {},
   setShowAddEditDomainModal: () => {},
-  setShowTagLinkModal: () => {},
 });
 
 export default function ModalProvider({ children }: { children: ReactNode }) {
@@ -34,7 +31,6 @@ export default function ModalProvider({ children }: { children: ReactNode }) {
     useAcceptInviteModal();
   const { setShowAddEditDomainModal, AddEditDomainModal } =
     useAddEditDomainModal({});
-  const { setShowTagLinkModal, TagLinkModal } = useTagLinkModal({});
 
   const { error, loading } = useProject();
   const { data: session } = useSession();
@@ -61,7 +57,6 @@ export default function ModalProvider({ children }: { children: ReactNode }) {
       value={{
         setShowAddProjectModal,
         setShowAddEditDomainModal,
-        setShowTagLinkModal,
       }}
     >
       <GoogleOauthModal />
@@ -70,7 +65,6 @@ export default function ModalProvider({ children }: { children: ReactNode }) {
         <AcceptInviteModal />
       )}
       <AddEditDomainModal />
-      <TagLinkModal />
       {children}
     </ModalContext.Provider>
   );
