+++ components/app/links/link-filters.tsx
@@ -8,13 +8,13 @@ import {
   useState,
 } from "react";
 import { nFormatter, setQueryString } from "#/lib/utils";
-import { ChevronRight, XCircle, Search } from "lucide-react";
+import { ChevronRight, XCircle, Search, Check, Trash } from "lucide-react";
 import useDomains from "#/lib/swr/use-domains";
 import { AnimatePresence, motion } from "framer-motion";
 import { SWIPE_REVEAL_ANIMATION_SETTINGS } from "#/lib/constants";
 import { useDebouncedCallback } from "use-debounce";
 import useLinks from "#/lib/swr/use-links";
-import { LoadingSpinner } from "#/ui/icons";
+import { LoadingCircle, LoadingSpinner } from "#/ui/icons";
 import useLinksCount from "#/lib/swr/use-links-count";
 import punycode from "punycode/";
 import Switch from "#/ui/switch";
@@ -22,8 +22,12 @@ import { useSession } from "next-auth/react";
 import { toast } from "sonner";
 import { ModalContext } from "#/ui/modal-provider";
 import useTags from "#/lib/swr/use-tags";
-import TagBadge from "@/components/app/links/tag-badge";
+import TagBadge, { COLORS_LIST } from "@/components/app/links/tag-badge";
 import { TagProps } from "#/lib/types";
+import { ThreeDots } from "@/components/shared/icons";
+import Popover from "@/components/shared/popover";
+import IconMenu from "@/components/shared/icon-menu";
+import { mutate } from "swr";
 
 export default function LinkFilters() {
   const { primaryDomain } = useDomains();
@@ -243,8 +247,7 @@ const TagsFilter = ({
   tagsCount: { tagId: string; _count: number }[];
 }) => {
   const router = useRouter();
-  const { slug } = router.query as { slug?: string };
-  const [collapsed, setCollapsed] = useState(true);
+  const [collapsed, setCollapsed] = useState(tags.length === 0 ? true : false);
   const [search, setSearch] = useState("");
   const [showMore, setShowMore] = useState(false);
 
@@ -312,7 +315,7 @@ const TagsFilter = ({
               .map(({ id, name, color, count }) => (
                 <div
                   key={id}
-                  className="relative flex cursor-pointer items-center space-x-3 rounded-md bg-gray-50 transition-all hover:bg-gray-100"
+                  className="group relative flex cursor-pointer items-center space-x-3 rounded-md bg-gray-50 transition-all hover:bg-gray-100"
                 >
                   <input
                     id={id}
@@ -329,7 +332,7 @@ const TagsFilter = ({
                     className="flex w-full cursor-pointer justify-between px-3 py-1.5 pl-0 text-sm font-medium text-gray-700"
                   >
                     <TagBadge name={name} color={color} />
-                    <p className="text-gray-500">{nFormatter(count)}</p>
+                    <TagPopover tag={{ id, name, color }} count={count} />
                   </label>
                 </div>
               ))}
@@ -345,6 +348,149 @@ const TagsFilter = ({
         )}
       </AnimatePresence>
     </fieldset>
+  );
+};
+
+const TagPopover = ({ tag, count }: { tag: TagProps; count: number }) => {
+  const router = useRouter();
+  const { slug } = router.query as { slug: string };
+  const [data, setData] = useState(tag);
+  const [openPopover, setOpenPopover] = useState(false);
+  const [processing, setProcessing] = useState(false);
+
+  const handleEdit = async (e) => {
+    e.stopPropagation();
+    e.preventDefault();
+    setProcessing(true);
+    fetch(`/api/projects/${slug}/tags/${tag.id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(data),
+    }).then((res) => {
+      setProcessing(false);
+      if (res.ok) {
+        toast.success("Tag updated");
+        mutate(`/api/projects/${slug}/tags`);
+      } else {
+        toast.error("Something went wrong");
+      }
+    });
+  };
+
+  const handleDelete = async () => {
+    setProcessing(true);
+    fetch(`/api/projects/${slug}/tags/${tag.id}`, {
+      method: "DELETE",
+    }).then((res) => {
+      setProcessing(false);
+      if (res.ok) {
+        toast.success("Tag deleted");
+        mutate(`/api/projects/${slug}/tags`);
+      } else {
+        toast.error("Something went wrong");
+      }
+    });
+  };
+
+  return processing ? (
+    <LoadingCircle />
+  ) : (
+    <Popover
+      content={
+        <div
+          data-exclude-click
+          className="flex w-48 flex-col divide-y divide-gray-200"
+        >
+          <div className="p-2">
+            <form
+              onClick={(e) => e.stopPropagation()} // prevent triggering <Command.Item> onClick
+              onKeyDown={(e) => e.stopPropagation()} // prevent triggering <Command.Item> onKeyDown
+              onSubmit={handleEdit}
+              className="relative py-1"
+            >
+              <div className="my-2 flex items-center justify-between px-3">
+                <p className="text-xs text-gray-500">Edit Tag</p>
+                {data !== tag && (
+                  <button className="text-xs text-gray-500">Save</button>
+                )}
+              </div>
+              <input
+                type="text"
+                autoFocus
+                required
+                onKeyDown={(e) => {
+                  // if ESC key pressed, close popover
+                  if (e.key === "Escape") {
+                    setOpenPopover(false);
+                  }
+                }}
+                value={data.name}
+                onChange={(e) => setData({ ...data, name: e.target.value })}
+                className="block w-full rounded-md border-gray-300 py-1 pr-7 text-sm text-gray-900 placeholder-gray-400 focus:border-gray-500 focus:outline-none focus:ring-gray-500"
+              />
+              <div className="grid grid-cols-3 gap-3 p-3 pb-0">
+                {COLORS_LIST.map(({ color, css }) => (
+                  <button
+                    key={color}
+                    type="button"
+                    className={`mx-auto flex h-6 w-6 items-center justify-center rounded-full transition-all duration-75 hover:scale-110 active:scale-90 ${css}`}
+                    onClick={(e) => {
+                      e.stopPropagation();
+                      setData({ ...data, color });
+                    }}
+                  >
+                    {data.color === color && <Check className="h-4 w-4" />}
+                  </button>
+                ))}
+              </div>
+            </form>
+          </div>
+          <div className="p-2">
+            <button
+              onClick={(e) => {
+                e.stopPropagation();
+                confirm(
+                  "Are you sure you want to delete this tag? All tagged links will be untagged, but they won't be deleted.",
+                ) && handleDelete();
+              }}
+              className="flex w-full items-center space-x-2 rounded-md p-2 text-red-600 transition-colors hover:bg-red-100 active:bg-red-200"
+            >
+              <IconMenu
+                text="Delete Tag"
+                icon={<Trash className="h-4 w-4 text-red-600" />}
+              />
+            </button>
+          </div>
+        </div>
+      }
+      align="end"
+      desktopOnly
+      openPopover={openPopover}
+      setOpenPopover={setOpenPopover}
+    >
+      <button
+        type="button"
+        onClick={() => setOpenPopover(!openPopover)}
+        className={`${
+          openPopover ? "bg-gray-200" : "hover:bg-gray-200"
+        } -mr-1 flex h-6 w-5 items-center justify-center rounded-md transition-colors`}
+      >
+        <ThreeDots
+          className={`h-4 w-4 text-gray-500 ${
+            openPopover ? "" : "hidden group-hover:block"
+          }`}
+        />
+        <p
+          className={`text-gray-500 ${
+            openPopover ? "hidden" : "group-hover:hidden"
+          }`}
+        >
+          {nFormatter(count)}
+        </p>
+      </button>
+    </Popover>
   );
 };
